sắp xếp và fix
