--- BE/app/Http/Controllers/AdminController/CommentController.php
@@ -1,78 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\AdminController;
-
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
-use App\Models\Comment;
-class CommentController extends Controller
-{
-    /**
-     * Display a listing of the resource.
-     */
-    public function index()
-    {
-        $comments = Comment::all();
-        return view('admin.comment.index', compact('comments'));
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        return view('admin.comment.create');
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(Request $request)
-    {
-        $request->validate([
-            'content' =>'required|string',
-        ]);
-        Comment::create($request->all());
-        return redirect()->route('admin.comment.index')->with('success', 'Comment created successfully');
-    }
-
-    /**
-     * Display the specified resource.
-     */
-    public function show(string $id)
-    {
-       //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-        $comment = Comment::findOrFail($id);
-        return view('admin.comment.edit', compact('comment'));
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, string $id)
-    {
-        $request->validate([
-            'content' =>'required|string',
-        ]);
-        $comment = Comment::findOrFail($id);
-        $comment->update($request->all());
-        return redirect()->route('admin.comment.index')->with('success', 'Comment updated successfully');
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(string $id)
-    {
-        $comment = Comment::findOrFail($id);
-        $comment->delete();
-        return redirect()->route('admin.comment.index')->with('success', 'Comment deleted successfully');
-    }
-}
