--- BE/app/Http/Controllers/AdminController/UserController.php
@@ -1,143 +1,0 @@
-<?php
-
-namespace App\Http\Controllers\AdminController;
-
-use App\Http\Controllers\Controller;
-use Illuminate\Http\Request;
-use App\Models\User;
-use Illuminate\Support\Facades\Hash;
-use Illuminate\Support\Str;
-use Illuminate\Support\Facades\Storage;
-
-class UserController extends Controller
-{
-    /**
-     * Display a listing of the resource.
-     */
-    public function index()
-    {
-        $users = User::all();
-        return view('admin.user.index', compact('users'));
-    }
-
-    /**
-     * Show the form for creating a new resource.
-     */
-    public function create()
-    {
-        return view('admin.user.create');
-    }
-
-    /**
-     * Store a newly created resource in storage.
-     */
-    public function store(Request $request)
-    {
-        $request->validate([
-            'name' => ['required', 'string', 'max:255'],
-            'email' => ['required', 'string', 'email', 'max:255', 'unique:users'],
-            'password' => ['required', 'string', 'min:8', 'confirmed'],
-            'type' => ['required', 'in:admin,user'],
-            'phone_number' => ['required', 'string', 'max:15'],
-            'address' => ['required', 'string', 'max:255'],
-            'gender' => ['nullable', 'in:male,female,other'],
-            'birthday' => ['nullable', 'date'],
-            'avatar' => ['nullable', 'image', 'mimes:jpeg,png,jpg,gif', 'max:2048'],
-        ]);
-
-        // Upload avatar nếu có
-        $avatarPath = null;
-        if ($request->hasFile('avatar')) {
-            $avatarPath = $request->file('avatar')->store('avatars', 'public');
-        }
-
-        User::create([
-            'name' => $request->name,
-            'email' => $request->email,
-            'password' => Hash::make($request->password),
-            'type' => $request->type,
-            'phone_number' => $request->phone_number,
-            'address' => $request->address,
-            'gender' => $request->gender,
-            'birthday' => $request->birthday,
-            'avatar' => $avatarPath,
-            'remember_token' => Str::random(10),
-        ]);
-
-        return redirect()->route('admin.user.index')->with('success', 'User created successfully');
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     */
-    public function edit(string $id)
-    {
-        $user = User::findOrFail($id);
-        return view('admin.user.edit', compact('user'));
-    }
-
-    /**
-     * Update the specified resource in storage.
-     */
-    public function update(Request $request, string $id)
-    {
-        $user = User::findOrFail($id);
-
-        $request->validate([
-            'name' => ['required', 'string', 'max:255'],
-            'email' => ['required', 'string', 'email', 'max:255', 'unique:users,email,' . $user->id],
-            'password' => ['nullable', 'string', 'min:8', 'confirmed'],
-            'type' => ['required', 'in:admin,user'],
-            'phone_number' => ['required', 'string', 'max:15'],
-            'address' => ['required', 'string', 'max:255'],
-            'gender' => ['nullable', 'in:male,female,other'],
-            'birthday' => ['nullable', 'date'],
-            'avatar' => ['nullable', 'image', 'mimes:jpeg,png,jpg,gif', 'max:2048'],
-        ]);
-
-        // Upload avatar nếu có
-        if ($request->hasFile('avatar')) {
-            // Xóa avatar cũ nếu có
-            if ($user->avatar) {
-                Storage::disk('public')->delete($user->avatar);
-            }
-            $avatarPath = $request->file('avatar')->store('avatars', 'public');
-            $user->avatar = $avatarPath;
-        }
-
-        // Cập nhật mật khẩu nếu có
-        if ($request->filled('password')) {
-            $user->password = Hash::make($request->password);
-        }
-
-        $user->update($request->except(['password', 'avatar']));
-
-        return redirect()->route('admin.user.index')->with('success', 'User updated successfully');
-    }
-
-    /**
-     * Remove the specified resource from storage.
-     */
-    public function destroy(string $id)
-    {
-        $user = User::findOrFail($id);
-
-        // Xóa avatar cũ nếu có
-        if ($user->avatar) {
-            Storage::disk('public')->delete($user->avatar);
-        }
-
-        $user->delete();
-        return redirect()->route('admin.user.index')->with('success', 'User deleted successfully');
-    }
-
-    public function login()
-    {
-        return view('admin.user.login');
-    }
-
-    public function forgotPassword()
-    {
-        return view('admin.user.forgot_password');
-    }
-}
